# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: HTTP Conflict 409
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    OntologySchema:
      type: object
      properties:
        ontologies:
          description: 'SchemaIds of ontologies, defining the terms of a schema'
          type: array
          items:
            type: string
            example: schemaId
        shapes:
          description: 'SchemaIds of shapes (SHACL Shapes Constraint Language) used to validate instances against the schema.'
          type: array
          items:
            type: string
            example: schemaId
        vocabularies:
          description: 'SchemaIds of controlled vocabularies used in the schema'
          type: array
          items:
            type: string
            example: schemaId
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /schemas:
    get:
      tags:
        - Schemas
      summary: 'Get the full list of ontologies, shapes and vocabularies.'
      description: 'The schema identifiers, returned by this endpoint can be used to download the schema files.'
      operationId: getSchemas
      responses:
        '200':
          description: 'References to ontologies, shapes and vocabularies.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologySchema'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'          
    post:
      tags:
        - Schemas
      summary: 'Add a new Schema to the catalogue.'
      description: 'This endpoint accepts either an ontology (OWL file), shape (SHACL file) or controlled vocabulary (SKOS file), which is added to the schemas of the catalogue.'
      operationId: addSchema
      security:
        - jwt: []
      requestBody:
        description: 'The file of the new schema. either an ontology (OWL file), shape (SHACL file) or controlled vocabulary (SKOS file).'
        content:
          application/rdf+xml:
            schema:
              type: string
              description: 'OWL file describing the ontology of the schema or SKOS file describing the controlled vocabulary.'
          application/ld+json:
            schema:
              type: string
              description: 'OWL or SHACL file describing the ontology.'
          application/json:
              schema:
                type: string
                description: 'OWL or SHACL file describing the ontology.'
          text/turtle:
            schema:
              type: string
              description: 'SHACL file describing the ontology of the schema.'
        required: true
      responses:
        '201':
          description: 'Created'
          headers:
            location:
              description: 'URL of the added Schema.'
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /schemas/{schemaId}:
    get:
      tags:
        - Schemas
      summary: Get a specific schema.
      operationId: getSchema
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'The schema for the given identifier. Depending on the type of the schema, either an ontology, shape graph or controlled vocabulary is returned.'
          content:
            application/rdf+xml:
              schema:
                type: string
                description: 'OWL file describing the ontology of the schema or SKOS file describing the controlled vocabulary.'
            text/turtle:
              schema:
                type: string
                description: 'SHACL file describing the ontology of the schema.'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Schemas
      summary: 'Replace a schema. This is only allowed for ontologies with the same IRI.'
      operationId: updateSchema
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema.'
          required: true
          schema:
            type: string
      security:
        - jwt: [ ]
      requestBody:
        description: 'The new ontology (OWL file).'
        content:
          application/rdf+xml:
            schema:
              type: string
              description: 'OWL file describing the ontology.'
          application/ld+json:
            schema:
              type: string
              description: 'OWL or SHACL file describing the ontology.'
          application/json:
            schema:
              type: string
              description: 'OWL or SHACL file describing the ontology.'
          text/turtle:
            schema:
              type: string
              description: 'SHACL file describing the ontology of the schema.'
        required: true
      responses:
        '200':
          description: 'Updated.'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Schemas
      summary: 'Delete a Schema'
      operationId: deleteSchema
      security:
        - jwt: [ ]
      parameters:
        - name: schemaId
          in: path
          description: 'Identifier of the Schema'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Deleted Schema'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'          
  /schemas/latest:
    get: 
      tags:
        - Schemas
      summary: 'Get the latest schema for a given type. If no term is specified, then the composite schema is returned.'
      operationId: getLatestSchema
      parameters:
        - name: type
          in: query
          description: 'Type of the schema.'
          required: true
          schema:
            type: string
            enum:
              - ontology
              - shape
              - vocabulary
        - name: term
          in: query
          description: 'The URI of the term of the requested Self-Description schema e.g. `http://w3id.org/gaia-x/service#ServiceOffering`'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'The latest schemas for the given type or term.'
          content:
            application/rdf+xml:
              schema:
                type: string
                description: 'OWL file describing the ontology of the schema or SKOS file describing the controlled vocabulary.'
            text/turtle:
              schema:
                type: string
                description: 'SHACL file describing the ontology of the schema.'
        '400':
          $ref: '#/components/responses/ClientError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...